[#27] refactor: PDF 버튼 공통 컴포넌트로 폴더 이동
